fix: correct typos in experimental natural language retriever module
