--- llama-index-experimental/llama_index/experimental/retrievers/natrual_language/__init__.py
@@ -1,7 +1,0 @@
-from llama_index.experimental.retrievers.natrual_language import (
-    nl_csv_retriever,
-    nl_data_frame_retierver,
-    nl_json_retriever,
-)
-
-__all__ = ["nl_data_frame_retierver", "nl_csv_retriever", "nl_json_retriever"]
